AI_Closet/requirements.txt

numpy>=1.26
pandas>=2.0
pillow>=10.0
opencv-python>=4.9
scikit-learn>=1.4
torch>=2.2
torchvision>=0.17
streamlit>=1.36
sqlalchemy>=2.0
python-dotenv>=1.0
matplotlib>=3.8
fastapi>=0.111
pydantic>=2.8
typing-extensions>=4.12
